Remove DynamicViewItem component
