+++ src/modules/form/components/viewable/DynamicViewFields.tsx
@@ -1,3 +1,4 @@
+import { Grid } from '@mui/material';
 import React, { ReactNode } from 'react';
 
 import {
@@ -11,7 +12,7 @@ import { isEnabled } from '../../util/formUtil';
 
 import DynamicViewGroup from './DynamicViewGroup';
 
-import DynamicViewItem from '@/modules/form/components/viewable/DynamicViewItem';
+import DynamicViewField from '@/modules/form/components/viewable/DynamicViewField';
 import {
   DisabledDisplay,
   FormDefinitionJson,
@@ -74,19 +75,30 @@ const DynamicViewFields: React.FC<Props> = ({
       );
     }
 
-    return (
-      <DynamicViewItem
-        key={item.linkId}
-        item={item}
-        isDisabled={isDisabled}
-        horizontal={horizontal}
-        nestingLevel={nestingLevel}
-        values={values}
-        renderFn={renderFn}
-        pickListArgs={pickListArgs}
-        itemChanged={itemChanged}
-      />
+    const itemComponent = (
+      <Grid item key={item.linkId}>
+        <DynamicViewField
+          item={item}
+          value={
+            isDisabled &&
+            item.disabledDisplay !== DisabledDisplay.ProtectedWithValue
+              ? undefined
+              : values[item.linkId]
+          }
+          disabled={isDisabled}
+          nestingLevel={nestingLevel}
+          horizontal={horizontal}
+          pickListArgs={pickListArgs}
+          // Needed because there are some enable/disabled and autofill dependencies that depend on PickListOption.labels that are fetched (PriorLivingSituation is an example)
+          adjustValue={itemChanged}
+          {...props}
+        />
+      </Grid>
     );
+    if (renderFn) {
+      return renderFn(itemComponent);
+    }
+    return itemComponent;
   };
 
   return <>{definition.item.map((item) => renderItem(item, 0))}</>;
